feat: add MemRaftLog, not finished

+++ server/src/main/java/com/github/dtprj/dongting/raft/impl/ReplicateManager.java
@@ -148,7 +148,7 @@ public class ReplicateManager {
 
     private void doReplicate(RaftMember member) {
         long nextIndex = member.getNextIndex();
-        int diff = (int) (raftStatus.getLastLogIndex() - nextIndex + 1);
+        long diff = raftStatus.getLastLogIndex() - nextIndex + 1;
         if (diff <= 0) {
             // no data to replicate
             return;
@@ -162,7 +162,7 @@ public class ReplicateManager {
             return;
         }
 
-        int limit = member.isMultiAppend() ? Math.min(rest, diff) : 1;
+        int limit = member.isMultiAppend() ? (int) Math.min(rest, diff) : 1;
 
         RaftTask first = raftStatus.getPendingRequests().get(nextIndex);
         RaftLog.LogIterator logIterator = member.getReplicateIterator();
